Вернул обработчики событий в editor.js + другие правки
