+++ js/render-thumbnails.js
@@ -1,13 +1,8 @@
 import {createThumbnail } from './create-thumbnail.js';
 import { openPictureModal, closePictureModal } from './render-modal.js';
 import { isEscKey } from './util.js';
-import { closeUploadOverlay } from './editor.js';
 
-// Значения атрибута name для текстовых полей
-const TEXT_FIELD_NAMES = ['hashtags', 'description'];
 const thumbnailsContainer = document.querySelector('.pictures');
-
-const isNotTextFields = (evt) => ! TEXT_FIELD_NAMES.includes(evt.target.name);
 
 // Обработчик нажатия на клавишу ESC на оверлее просмотра большого изображения
 const onModalEscKeydown = (evt) => {
@@ -21,20 +16,6 @@ const onModalEscKeydown = (evt) => {
 const onModalCloseButtonClick = (evt) => {
   evt.preventDefault();
   closePictureModal();
-};
-
-// Обработчик клика по кнопке закрытия оверлея редактирования загружаемого изображения
-const onImgUploadOverlayCancelButtonClick = (evt) => {
-  evt.preventDefault();
-  closeUploadOverlay();
-};
-
-// Обработчик нажатия на клавишу ESC на оверлее редактирования загружаемого изображения
-const onImgUploadOverlayEscKeydown = (evt) => {
-  if (isEscKey(evt) && isNotTextFields(evt)) {
-    evt.preventDefault();
-    closeUploadOverlay();
-  }
 };
 
 // Функция создания одной миниатюры
@@ -53,4 +34,4 @@ const renderThumbnails = (photos) => {
   thumbnailsContainer.append(...photos.map(renderThumbnail));
 };
 
-export {renderThumbnails, onModalCloseButtonClick, onModalEscKeydown, onImgUploadOverlayCancelButtonClick, onImgUploadOverlayEscKeydown};
+export { renderThumbnails, onModalCloseButtonClick, onModalEscKeydown };
