+++ js/util.js
@@ -1,9 +1,10 @@
-const KEY_VALUES = ['Escape', 'Esc'];
+const KEYBOARD_KEYS = ['Escape', 'Esc'];
 // Функция для проверки строки на максимальную длину
 const checkTextLength = (text, maxLength) => text.length <= maxLength;
 // Проверка что нажатая клавиша - ESC
-const isEscKey = (evt) => KEY_VALUES.includes(evt.key);
+const isEscKey = (evt) => KEYBOARD_KEYS.includes(evt.key);
 // Функция проверки массива на уникальность
-const isArrayUnique = (elements) => !elements.some((element, index) => elements.indexOf(element) !== index);
+const isArrayUnique = (items) => !items.some((item, index) => items.indexOf(item, index + 1) !== -1);
+
 
 export { checkTextLength, isEscKey, isArrayUnique };
