+++ js/validate.js
@@ -2,12 +2,12 @@ import { checkTextLength, isArrayUnique } from './util.js';
 
 // Максимальная длинна комментария
 const COMMENT_MAX_LENGTH = 140;
-
-// Перечисление настроек хэштегов
-const Hashtags = {
-  AMOUNT: 5,
-  MIN_LENGTH: 2,
-  MAX_LENGTH: 20
+// Максимальное количество хэштегов
+const MAX_HASHTAG = 5;
+// Длина хэштега
+const HashtagLength = {
+  MIN: 2,
+  MAX: 20
 };
 
 // Форма загрузки изображения на сайт
@@ -19,6 +19,9 @@ const commentInput = imgUploadOverlay.querySelector('.text__description');
 // Поле ввода хэштега
 const hashtagsInput = imgUploadOverlay.querySelector('.text__hashtags');
 
+const hashtagEndingRegExp = /[^-_=+;:,.]$/m;
+const hashtagRegExp = /^#[a-zA-Zа-яА-ЯёЁ0-9]{0,}$/;
+
 const pristine = new Pristine (uploadForm, {
   classTo: 'img-upload__field-wrapper',
   errorClass: 'form-item__invalid',
@@ -29,26 +32,26 @@ const pristine = new Pristine (uploadForm, {
 });
 
 // Функция очистки от лишних пробелов, приведения к нижнему регистру и помещения в массив введённых хэштегов
-const prepareHashtags = (value) => value !== '' ? value.trim().toLowerCase().split(' ') : [];
+const parseHashtagsInput = (value) => value !== '' ? value.trim().toLowerCase().split(' ') : [];
 
 // Проверки комментариев и хэштегов на валидность
 
-pristine.addValidator(commentInput, (comment) => checkTextLength(comment, COMMENT_MAX_LENGTH),
+pristine.addValidator(commentInput, (value) => checkTextLength(value, COMMENT_MAX_LENGTH),
   'Комментарий не должен быть длиннее 140 символов');
 
-pristine.addValidator(hashtagsInput, (hashtags) => prepareHashtags(hashtags).length <= Hashtags.AMOUNT,
+pristine.addValidator(hashtagsInput, (value) => parseHashtagsInput(value).length <= MAX_HASHTAG,
   'Не больше 5 хэштегов');
 
-pristine.addValidator(hashtagsInput, (hashtags) => hashtags === '' || prepareHashtags(hashtags).every((value) => /[^-_=+;:,.]$/m.test(value)),
+pristine.addValidator(hashtagsInput, (value) => value === '' || parseHashtagsInput(value).every((item) => hashtagEndingRegExp.test(item)),
   'Хэштеги разделяются пробелом');
 
-pristine.addValidator(hashtagsInput, (hashtags) => isArrayUnique(prepareHashtags(hashtags)),
+pristine.addValidator(hashtagsInput, (value) => isArrayUnique(parseHashtagsInput(value)),
   'Хэштеги не должны повторяться');
 
-pristine.addValidator(hashtagsInput, (hashtags) => hashtags === '' || prepareHashtags(hashtags).every((value) => /^#[a-zA-Zа-яА-ЯёЁ0-9]{0,}$/.test(value)),
+pristine.addValidator(hashtagsInput, (value) => value === '' || parseHashtagsInput(value).every((item) => hashtagRegExp.test(item)),
   'Хэштег начинается с # и состоит только из букв и цифр');
 
-pristine.addValidator(hashtagsInput, (hashtags) => hashtags === '' || prepareHashtags(hashtags).every((value) => value.length >= Hashtags.MIN_LENGTH && value.length <= Hashtags.MAX_LENGTH),
+pristine.addValidator(hashtagsInput, (value) => value === '' || parseHashtagsInput(value).every((item) => item.length >= HashtagLength.MIN && item.length <= HashtagLength.MAX),
   'Длина хэштега — от 1 до 19 символов после #');
 
 // Обработчик действия при отправке формы
@@ -60,7 +63,7 @@ uploadForm.addEventListener('submit', (evt) => {
 });
 
 /* Обработчик события изменения формы чтобы при закрытии попапа
-   очищались предупреждений от старых проверок */
+   очищались предупреждения от старых проверок */
 uploadForm.addEventListener('change', () => {
   pristine.validate();
 });
